+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2.0.0 under development
 
-- Enh #289: Implement `SqlParser` and `ExpressionBuilder` driver classes (@Tigrov)
+- Enh #289, #352: Implement and use `SqlParser` class (@Tigrov)
 - New #273: Implement `ColumnSchemaInterface` classes according to the data type of database table columns
   for type casting performance. Related with yiisoft/db#752 (@Tigrov)
 - Chg #307: Replace call of `SchemaInterface::getRawTableName()` to `QuoterInterface::getRawTableName()` (@Tigrov)